rack session fix
