+++ app/controllers/users/registrations_controller.rb
@@ -1,47 +1,18 @@
 class Users::RegistrationsController < Devise::RegistrationsController
+  include RackSessionsFix
   respond_to :json
   private
 
-  def respond_with(resource, _opts = {})
-    if request.method == "POST" && resource.persisted?
+  def respond_with(current_user, _opts = {})
+    if resource.persisted?
       render json: {
-        status: {code: 200, message: "Signed up sucessfully."},
-        data: UserSerializer.new(resource).serializable_hash[:data][:attributes]
-      }, status: :ok
-    elsif request.method == "DELETE"
-      render json: {
-        status: { code: 200, message: "Account deleted successfully."}
-      }, status: :ok
+        status: {code: 200, message: 'Signed up successfully.'},
+        data: UserSerializer.new(current_user).serializable_hash[:data][:attributes]
+      }
     else
       render json: {
-        status: {code: 422, message: "User couldn't be created successfully. #{resource.errors.full_messages.to_sentence}"}
+        status: {message: "User couldn't be created successfully. #{current_user.errors.full_messages.to_sentence}"}
       }, status: :unprocessable_entity
-    end
-  end
-end
-
-class Users::SessionsController < Devise::SessionsController
-  respond_to :json
-  private
-
-  def respond_with(resource, _opts = {})
-    render json: {
-      status: {code: 200, message: 'Logged in sucessfully.'},
-      data: UserSerializer.new(resource).serializable_hash[:data][:attributes]
-    }, status: :ok
-  end
-
-  def respond_to_on_destroy
-    if current_user
-      render json: {
-        status: 200,
-        message: "logged out successfully"
-      }, status: :ok
-    else
-      render json: {
-        status: 401,
-        message: "Couldn't find an active session."
-      }, status: :unauthorized
     end
   end
 end
